Pass two of fixing `make build` on MacOS
